+++ cloud-provider-payment-8001/src/main/java/my/lsh/cloud/service/PaymentServiceI.java
@@ -6,5 +6,5 @@ public interface PaymentServiceI {
 
     int create(PaymentEntity paymentEntity);
 
-    PaymentEntity getPaymentById(Integer id);
+    PaymentEntity getPaymentById(Long id);
 }
