+++ cloud-provider-payment-8001/src/main/java/my/lsh/cloud/service/impl/PaymentServiceImpl.java
@@ -16,7 +16,7 @@ public class PaymentServiceImpl implements PaymentServiceI {
     }
 
     @Override
-    public PaymentEntity getPaymentById(Integer id) {
+    public PaymentEntity getPaymentById(Long id) {
         return dao.getPaymentById(id);
     }
 }
